Refactor docker service fixtures
