+++ tests/unit/test_channels/test_plugin.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import sys
 from secrets import token_hex
 from typing import cast
 from unittest.mock import AsyncMock, MagicMock
@@ -19,11 +18,6 @@ from litestar.testing import TestClient, create_test_client
 from litestar.types.asgi_types import WebSocketMode
 
 from .util import get_from_stream
-
-pytestmark = [
-    pytest.mark.usefixtures("redis_service"),
-    pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
-]
 
 
 @pytest.fixture
