+++ tests/unit/test_contrib/test_sqlalchemy/test_repository/helpers.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
 import inspect
-import sys
 from datetime import datetime, timezone
 from typing import Any, Awaitable, TypeVar, cast, overload
-
-import pytest
 
 T = TypeVar("T")
 
@@ -38,9 +35,3 @@ def update_raw_records(raw_authors: list[dict[str, Any]], raw_rules: list[dict[s
     for raw_rule in raw_rules:
         raw_rule["created_at"] = datetime.strptime(raw_rule["created_at"], "%Y-%m-%dT%H:%M:%S").astimezone(timezone.utc)
         raw_rule["updated_at"] = datetime.strptime(raw_rule["updated_at"], "%Y-%m-%dT%H:%M:%S").astimezone(timezone.utc)
-
-
-mark_requires_docker = [
-    pytest.mark.sqlalchemy_integration,
-    pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
-]
