+++ tests/unit/test_contrib/test_sqlalchemy/test_repository/test_repo_uuid.py
@@ -31,16 +31,16 @@ from tests.unit.test_contrib.test_sqlalchemy.models_uuid import (
     UUIDRule,
 )
 
-from .helpers import mark_requires_docker, maybe_async, update_raw_records
+from .helpers import maybe_async, update_raw_records
 
 
 @pytest.fixture(
     params=[
         pytest.param("sqlite_engine", marks=pytest.mark.sqlalchemy_sqlite),
-        pytest.param("duckdb_engine", marks=[pytest.mark.sqlalchemy_duckdb, *mark_requires_docker]),
-        pytest.param("oracle_engine", marks=[pytest.mark.sqlalchemy_oracledb, *mark_requires_docker]),
-        pytest.param("psycopg_engine", marks=[pytest.mark.sqlalchemy_psycopg_sync, *mark_requires_docker]),
-        pytest.param("spanner_engine", marks=[pytest.mark.sqlalchemy_spanner, *mark_requires_docker]),
+        pytest.param("duckdb_engine", marks=[pytest.mark.sqlalchemy_duckdb, pytest.mark.sqlalchemy_integration]),
+        pytest.param("oracle_engine", marks=[pytest.mark.sqlalchemy_oracledb, pytest.mark.sqlalchemy_integration]),
+        pytest.param("psycopg_engine", marks=[pytest.mark.sqlalchemy_psycopg_sync, pytest.mark.sqlalchemy_integration]),
+        pytest.param("spanner_engine", marks=[pytest.mark.sqlalchemy_spanner, pytest.mark.sqlalchemy_integration]),
     ]
 )
 def engine(request: FixtureRequest) -> Engine:
