+++ tests/unit/test_contrib/test_sqlalchemy/test_repository/test_sqlalchemy_oracledb_json.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 import platform
-import sys
 
 import pytest
 from sqlalchemy import Engine
@@ -12,9 +11,7 @@ from sqlalchemy.schema import CreateTable
 from tests.unit.test_contrib.test_sqlalchemy.models_uuid import UUIDEventLog
 
 pytestmark = [
-    pytest.mark.skipif(sys.platform != "linux", reason="docker not available on this platform"),
     pytest.mark.skipif(platform.uname()[4] != "x86_64", reason="oracle not available on this platform"),
-    pytest.mark.sqlalchemy_integration,
 ]
